requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
matplotlib>=3.4.0
scikit-learn>=1.0.0
PyYAML>=6.0